More extractCurrentPageFromUrl + tests into url_utils
