+++ packages/devtools_app/lib/src/analytics/_analytics_web.dart
@@ -17,6 +17,7 @@ import '../app.dart';
 import '../config_specific/logger/logger.dart';
 import '../config_specific/server/server.dart' as server;
 import '../config_specific/url/url.dart';
+import '../primitives/url_utils.dart';
 import '../screens/performance/performance_screen.dart';
 import '../screens/profiler/profiler_screen.dart';
 import '../shared/globals.dart';
